Add files via upload
Our command line. Currently the classes and methods you can call are all in this file.

It can be tested by entering "Check empid" which will return that the ID is valid. The next step is to add the ability to have a variable sent with the command, then the ability to call classes from other files.